Update models.py with IntegerField

+++ chord_metadata_service/mohpackets/models.py
@@ -91,7 +91,7 @@ class PrimaryDiagnosis(models.Model):
         max_length=128, null=True, blank=True
     )
     lymph_nodes_examined_method = models.CharField(max_length=64, null=True, blank=True)
-    number_lymph_nodes_positive = models.PositiveSmallIntegerField(
+    number_lymph_nodes_positive = models.IntegerField(
         null=True, blank=True
     )
     clinical_tumour_staging_system = models.CharField(
@@ -210,8 +210,8 @@ class Treatment(models.Model):
     treatment_end_date = models.JSONField(null=True, blank=True)
     treatment_setting = models.CharField(max_length=128, null=True, blank=True)
     treatment_intent = models.CharField(max_length=128, null=True, blank=True)
-    days_per_cycle = models.PositiveSmallIntegerField(null=True, blank=True)
-    number_of_cycles = models.PositiveSmallIntegerField(null=True, blank=True)
+    days_per_cycle = models.IntegerField(null=True, blank=True)
+    number_of_cycles = models.IntegerField(null=True, blank=True)
     response_to_treatment_criteria_method = models.CharField(
         max_length=255, null=True, blank=True
     )
@@ -245,10 +245,10 @@ class Chemotherapy(models.Model):
     chemotherapy_drug_dose_units = models.CharField(
         max_length=64, null=True, blank=True
     )
-    prescribed_cumulative_drug_dose = models.PositiveSmallIntegerField(
+    prescribed_cumulative_drug_dose = models.IntegerField(
         blank=True, null=True
     )
-    actual_cumulative_drug_dose = models.PositiveSmallIntegerField(
+    actual_cumulative_drug_dose = models.IntegerField(
         blank=True, null=True
     )
 
@@ -276,10 +276,10 @@ class HormoneTherapy(models.Model):
     drug_name = models.CharField(max_length=255, null=True, blank=True)
     drug_reference_identifier = models.CharField(max_length=64, null=True, blank=True)
     hormone_drug_dose_units = models.CharField(max_length=64, null=True, blank=True)
-    prescribed_cumulative_drug_dose = models.PositiveSmallIntegerField(
+    prescribed_cumulative_drug_dose = models.IntegerField(
         blank=True, null=True
     )
-    actual_cumulative_drug_dose = models.PositiveSmallIntegerField(
+    actual_cumulative_drug_dose = models.IntegerField(
         blank=True, null=True
     )
 
@@ -305,10 +305,10 @@ class Radiation(models.Model):
     submitter_treatment_id = models.CharField(max_length=64, null=False, blank=False)
     radiation_therapy_modality = models.CharField(max_length=255, null=True, blank=True)
     radiation_therapy_type = models.CharField(max_length=64, null=True, blank=True)
-    radiation_therapy_fractions = models.PositiveSmallIntegerField(
+    radiation_therapy_fractions = models.IntegerField(
         null=True, blank=True
     )
-    radiation_therapy_dosage = models.PositiveSmallIntegerField(null=True, blank=True)
+    radiation_therapy_dosage = models.IntegerField(null=True, blank=True)
     anatomical_site_irradiated = models.CharField(max_length=255, null=True, blank=True)
     radiation_boost = models.BooleanField(blank=True, null=True)
     reference_radiation_treatment_id = models.CharField(
@@ -342,10 +342,10 @@ class Immunotherapy(models.Model):
     immunotherapy_drug_dose_units = models.CharField(
         max_length=64, null=True, blank=True
     )
-    prescribed_cumulative_drug_dose = models.PositiveSmallIntegerField(
+    prescribed_cumulative_drug_dose = models.IntegerField(
         blank=True, null=True
     )
-    actual_cumulative_drug_dose = models.PositiveSmallIntegerField(
+    actual_cumulative_drug_dose = models.IntegerField(
         blank=True, null=True
     )
 
@@ -373,9 +373,9 @@ class Surgery(models.Model):
     surgery_type = models.CharField(max_length=255, null=True, blank=True)
     surgery_site = models.CharField(max_length=255, null=True, blank=True)
     surgery_location = models.CharField(max_length=128, null=True, blank=True)
-    tumour_length = models.PositiveSmallIntegerField(null=True, blank=True)
-    tumour_width = models.PositiveSmallIntegerField(null=True, blank=True)
-    greatest_dimension_tumour = models.PositiveSmallIntegerField(null=True, blank=True)
+    tumour_length = models.IntegerField(null=True, blank=True)
+    tumour_width = models.IntegerField(null=True, blank=True)
+    greatest_dimension_tumour = models.IntegerField(null=True, blank=True)
     tumour_focality = models.CharField(max_length=64, null=True, blank=True)
     residual_tumour_classification = models.CharField(
         max_length=64, null=True, blank=True
@@ -461,9 +461,9 @@ class Biomarker(models.Model):
     submitter_treatment_id = models.CharField(max_length=64, null=True, blank=True)
     submitter_follow_up_id = models.CharField(max_length=64, null=True, blank=True)
     test_date = models.JSONField(null=True, blank=True)
-    psa_level = models.PositiveSmallIntegerField(null=True, blank=True)
-    ca125 = models.PositiveSmallIntegerField(null=True, blank=True)
-    cea = models.PositiveSmallIntegerField(null=True, blank=True)
+    psa_level = models.IntegerField(null=True, blank=True)
+    ca125 = models.IntegerField(null=True, blank=True)
+    cea = models.IntegerField(null=True, blank=True)
     er_status = models.CharField(max_length=64, null=True, blank=True)
     er_percent_positive = models.FloatField(null=True, blank=True)
     pr_status = models.CharField(max_length=64, null=True, blank=True)
@@ -496,7 +496,7 @@ class Comorbidity(models.Model):
     laterality_of_prior_malignancy = models.CharField(
         max_length=64, null=True, blank=True
     )
-    age_at_comorbidity_diagnosis = models.PositiveSmallIntegerField(
+    age_at_comorbidity_diagnosis = models.IntegerField(
         null=True, blank=True
     )
     comorbidity_type_code = models.CharField(max_length=64, null=True, blank=True)
